Moved update-check files to services folder
fix https://linear.app/ghost/issue/ENG-2038/move-core-update-check-to-service-folder

- these files essentially form a service, but they were straggling
  outside the services folder
- this moves them into the services folder so we can start to structure
  the code better
